Move related methods into a separate fixture
